Mockup
Geometry was coded as an example, the transparency =  blockchain transaction. You learn as you go and hear music, learn about the game, which is the blockchain. It is endless. The (i) is where the alpha is dropped, basically all things Blockchain and NFT's. See the grid, that's how simple the explanation of it is. Great for anyone who wants to learn. Has the option to turn off audio or visuals (temporary visuals, we don't want to give anyone a seizure of course).